Update entities datatypes

+++ MyTennis.DAL/Entities/League.cs
@@ -2,7 +2,7 @@
 {
     class League
     {
-        public int Id { get; set; }
+        public byte Id { get; set; }
         public string Name { get; set; }
     }
 }
