+++ MyTennis.DAL/Entities/Result.cs
@@ -4,8 +4,8 @@
     {
         public int Id { get; set; }
         public Game GameId { get; set; }
-        public int SetNr { get; set; }
-        public int ScoreTeamMember { get; set; }
-        public int ScoreOpponent { get; set; }
+        public byte SetNr { get; set; }
+        public byte ScoreTeamMember { get; set; }
+        public byte ScoreOpponent { get; set; }
     }
 }
